requirements.txt
albumentations
certifi
cycler
decorator
imageio
imgaug
kiwisolver
matplotlib
networkx
numpy
opencv-python
opencv-python-headless
Pillow
pyparsing
python-dateutil
PyWavelets
PyYAML
scikit-image
scipy
six
torch
torchvision
